To divided main file on models.cpp and initlayer.cpp
